Move `service-worker` to `internal/service-worker`
