+++ src/internal/service-worker/service-worker.ts
@@ -2,7 +2,7 @@
 /// <reference lib="esnext" />
 /// <reference lib="webworker" />
 
-import * as logging from "../pkg/logging";
+import * as logging from "../../pkg/logging";
 
 const log = logging.setDefaultLogger(
   new logging.Logger("Service-Worker", logging.Level.Debug),
