+++ vite.config.ts
@@ -6,7 +6,7 @@ import { buildSync } from "esbuild";
 
 const serviceWorkerSrc = join(
   process.cwd(),
-  "src/service-worker/service-worker.ts",
+  "src/internal/service-worker/service-worker.ts",
 );
 
 const serviceWorkerDest = join(process.cwd(), "dist", "service-worker.js");
